Move files into a "src" folder
