+++ package.json
@@ -2,8 +2,8 @@
   "name": "discord-cross-hosting",
   "version": "2.1.8",
   "description": "The first package, which allows broadcastEvaling over Cross Hosted Machines and effecient Machine & Shard Managing.",
-  "main": "index.js",
-  "types": "./typings.d.ts",
+  "main": "./src/index.js",
+  "types": "./typings/index.d.ts",
   "dependencies": {
     "net-ipc": "github:timotejroiko/net-ipc",
     "node-fetch": "^2.6.1"
