+++ src/Managers/Client.js
