+++ src/Managers/Shard.js
